Fix CNEOWeaponInfo::Parse
Whoopsie

+++ mp/src/game/shared/NeotokyoSource/neo_weapon_parse.cpp
@@ -13,11 +13,9 @@ FileWeaponInfo_t* CreateWeaponInfo()
 	return new CNEOWeaponInfo;
 }
 
-
 CNEOWeaponInfo::CNEOWeaponInfo()
 {
 }
-
 
 void CNEOWeaponInfo::Parse( KeyValues *pKeyValuesData, const char *szWeaponName )
 {
@@ -31,27 +29,27 @@ void CNEOWeaponInfo::Parse( KeyValues *pKeyValuesData, const char *szWeaponName 
 	m_iAimType = pKeyValuesData->GetInt( "AimType", 1 );
 	m_iScopeStyle = pKeyValuesData->GetInt( "ScopeStyle", 0 );	
 
-	m_fVMFov = pKeyValuesData->GetInt( "VMFov", 54.f );
-	m_fVMAimFov = pKeyValuesData->GetInt( "VMAimFov", 54.f );
-	m_fAimFov = pKeyValuesData->GetInt( "AimFov", 54.f );
+	m_fVMFov = pKeyValuesData->GetFloat( "VMFov", 54.f );
+	m_fVMAimFov = pKeyValuesData->GetFloat( "VMAimFov", 54.f );
+	m_fAimFov = pKeyValuesData->GetFloat( "AimFov", 54.f );
 
-	m_fVMOffsetUp = pKeyValuesData->GetInt( "VMOffsetUp", 0.f );
-	m_fVMOffsetForward = pKeyValuesData->GetInt( "VMOffsetForward", 0.f );
-	m_fVMOffsetRight = pKeyValuesData->GetInt( "VMOffsetRight", 0.f );
+	m_fVMOffsetUp = pKeyValuesData->GetFloat( "VMOffsetUp", 0.f );
+	m_fVMOffsetForward = pKeyValuesData->GetFloat( "VMOffsetForward", 0.f );
+	m_fVMOffsetRight = pKeyValuesData->GetFloat( "VMOffsetRight", 0.f );
 
-	m_fVMAimOffsetUp = pKeyValuesData->GetInt( "VMAimOffsetUp", 0.f );
-	m_fVMAimOffsetForward = pKeyValuesData->GetInt( "VMAimOffsetForward", 0.f );
-	m_fVMAimOffsetRight = pKeyValuesData->GetInt( "VMAimOffsetRight", 0.f );
+	m_fVMAimOffsetUp = pKeyValuesData->GetFloat( "VMAimOffsetUp", 0.f );
+	m_fVMAimOffsetForward = pKeyValuesData->GetFloat( "VMAimOffsetForward", 0.f );
+	m_fVMAimOffsetRight = pKeyValuesData->GetFloat( "VMAimOffsetRight", 0.f );
 
-	m_fVMAngleYaw = pKeyValuesData->GetInt( "VMAngleYaw", 0.f );
-	m_fVMAnglePitch = pKeyValuesData->GetInt( "VMAnglePitch", 0.f );
-	m_fVMAngleRoll = pKeyValuesData->GetInt( "VMAngleRoll", 0.f );
+	m_fVMAngleYaw = pKeyValuesData->GetFloat( "VMAngleYaw", 0.f );
+	m_fVMAnglePitch = pKeyValuesData->GetFloat( "VMAnglePitch", 0.f );
+	m_fVMAngleRoll = pKeyValuesData->GetFloat( "VMAngleRoll", 0.f );
 
-	m_fVMAimAngleYaw = pKeyValuesData->GetInt( "VMAimAngleYaw", 0.f );
-	m_fVMAimAnglePitch = pKeyValuesData->GetInt( "VMAimAnglePitch", 0.f );
-	m_fVMAimAngleRoll = pKeyValuesData->GetInt( "VMAimAngleRoll", 0.f );
+	m_fVMAimAngleYaw = pKeyValuesData->GetFloat( "VMAimAngleYaw", 0.f );
+	m_fVMAimAnglePitch = pKeyValuesData->GetFloat( "VMAimAnglePitch", 0.f );
+	m_fVMAimAngleRoll = pKeyValuesData->GetFloat( "VMAimAngleRoll", 0.f );
 
-	m_fPenetration = pKeyValuesData->GetInt( "Penetration", 1.f );
+	m_fPenetration = pKeyValuesData->GetFloat( "Penetration", 1.f );
 
 	// These were always in here
 	Q_strncpy( m_szAnim_prefix, pKeyValuesData->GetString( "anim_prefix" ), MAX_WEAPON_PREFIX );
@@ -59,8 +57,8 @@ void CNEOWeaponInfo::Parse( KeyValues *pKeyValuesData, const char *szWeaponName 
 	m_iDamage = pKeyValuesData->GetInt( "Damage", 42 );
 	m_iBullets = pKeyValuesData->GetInt( "Bullets", 1 );
 
-	m_fCycleTime = pKeyValuesData->GetInt( "CycleTime", 0.15f );
-	m_fTPMuzzleFlashScale = pKeyValuesData->GetInt( "TPMuzzleFlashScale", 1.0f );
+	m_fCycleTime = pKeyValuesData->GetFloat( "CycleTime", 0.15f );
+	m_fTPMuzzleFlashScale = pKeyValuesData->GetFloat( "TPMuzzleFlashScale", 1.0f );
 }
 
 // Check the header for more info -Ochii
